Banner fetching moved to find-games.vue & Some struc improvements
